+++ src/common/core/modsecurity/files/coreruleset/util/join-multiline-rules/join.py
@@ -20,16 +20,16 @@
 #
 # Example:
 #
-#   SecRule &TX:paranoia_level "@eq 0" \
+#   SecRule &TX:BLOCKING_PARANOIA_LEVEL "@eq 0" \
 #      "id:901120,\
 #      phase:1,\
 #      pass,\
 #      nolog,\
-#      setvar:tx.paranoia_level=1"
+#      setvar:tx.blocking_paranoia_level=1"
 #
 # will be outputted as:
 #
-#   SecRule &TX:paranoia_level "@eq 0" "id:901120,phase:1,pass,nolog,setvar:tx.paranoia_level=1"
+#   SecRule &TX:BLOCKING_PARANOIA_LEVEL "@eq 0" "id:901120,phase:1,pass,nolog,setvar:tx.blocking_paranoia_level=1"
 #
 
 import fileinput, sys